requirements.txt
lightning-bolts==0.5.0
numpy==1.22.4
opencv-python==4.5.5.62
Pillow==9.0.0
pycocotools==2.0.5
scikit-image==0.19.1
tensorboard==2.10.1
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.62.3